+++ botfront/imports/ui/components/templates/mutations.js
@@ -1,5 +1,12 @@
 import gql from 'graphql-tag';
 
+export const DELETE_BOT_RESPONSE = gql`
+mutation deleteResponse($projectId: String!, $key: String!) {
+    deleteResponse(projectId: $projectId, key: $key){
+      success
+    }
+}`;
+
 export const CREATE_BOT_RESPONSE = gql`
 mutation createResponse($projectId: String!, $response: BotResponseInput) {
     createResponse(projectId: $projectId, response: $response){
