+++ botfront/imports/ui/components/templates/queries.js
