--- botfront/imports/ui/layouts/graphQL/subscriptions.js
@@ -1,8 +1,0 @@
-import gql from 'graphql-tag';
-
-export const RESPONSE_ADDED = gql`
-subscription newResponse($projectId: String!) {
-    botResponseAdded(projectId: $projectId) {
-        key
-    }
-}`;
