+++ botfront/imports/ui/layouts/graphql.js
@@ -1,5 +1,12 @@
 import gql from 'graphql-tag';
 
+export const RESPONSE_ADDED = gql`
+subscription newResponse($projectId: String!) {
+    botResponseAdded(projectId: $projectId) {
+        key
+    }
+}`;
+
 export const GET_BOT_RESPONSES = gql`
 query retreiveBotResponses($projectId: String!) {
     botResponses(projectId: $projectId) {
@@ -18,5 +25,19 @@ query getResponse($template: StringOrListOfStrings!, $projectId: String!, $langu
         ...on QuickReplyPayload { buttons { title, type, ...on WebUrlButton { url } ...on PostbackButton { payload } } }
         ...on ImagePayload { image }
         ...on CustomPayload { buttons { title, type, ...on WebUrlButton { url } ...on PostbackButton { payload } }, elements, attachment, image, custom }
+    }
+}`;
+
+export const CREATE_BOT_RESPONSE = gql`
+mutation createResponse($projectId: String!, $response: BotResponseInput) {
+    createResponse(projectId: $projectId, response: $response){
+        success
+    }
+}`;
+
+export const UPDATE_BOT_RESPONSE = gql`
+mutation updateResponse($projectId: String!, $_id: String!, $response: BotResponseInput) {
+    updateResponse(projectId: $projectId, _id: $_id, response: $response){
+        success
     }
 }`;
